Consolidate String extensions
