+++ WordPress/Classes/Extensions/Media+Sync.swift
@@ -1,4 +1,5 @@
 import Foundation
+import CoreData
 
 extension Media {
