+++ WordPress/Classes/Extensions/Post+BloggingPrompts.swift
@@ -1,3 +1,5 @@
+import Foundation
+
 extension Post {
 
     func prepareForPrompt(_ prompt: BloggingPrompt?) {
