+++ WordPress/Classes/Extensions/String+Helpers.swift
@@ -30,4 +30,23 @@ extension String {
             .replacingOccurrences(of: "[\n]{3,}", with: "\n\n", options: .regularExpression, range: nil)
             .replacingOccurrences(of: "\\s+", with: " ", options: .regularExpression, range: nil)
     }
+
+    var replacingLastSpaceWithNonBreakingSpace: String {
+        if let lastSpace = range(of: " ", options: .backwards, locale: .current) {
+            return replacingCharacters(in: lastSpace, with: "\u{00a0}")
+        }
+        return self
+    }
+
+    /// Trims the trailing characters from the string to ensure the resulting string doesn't exceed the provided limit.
+    /// If the string is equal to or shorter than the limit, the string is returned without modifications
+    /// If the string is longer, the trailing characters are trimmed and replaced with an ellipsis character,
+    /// ensuring the length is equal to the limit
+    func truncate(with limit: Int) -> String {
+        guard count > limit else {
+            return self
+        }
+        let prefix = self.prefix(limit - 1)
+        return "\(prefix)…"
+    }
 }
